Added scripts + rxp node

+++ contracts/src/IOpenRankManager.sol
@@ -28,14 +28,14 @@ interface IOpenRankManager {
     struct MetaChallenge {
         address challenger;
         uint256 computeId;
-        uint256 subJobId;
+        uint32 subJobId;
         uint256 timestamp;
         uint256 requestIndex;
     }
 
     event MetaComputeRequestEvent(uint256 indexed computeId, bytes32 jobDescriptionId);
     event MetaComputeResultEvent(uint256 indexed computeId, bytes32 commitment, bytes32 resultsId);
-    event MetaChallengeEvent(uint256 indexed computeId, uint256 subJobId);
+    event MetaChallengeEvent(uint256 indexed computeId, uint32 subJobId);
 
     function setImageId(
         uint32 _imageId
@@ -48,7 +48,7 @@ interface IOpenRankManager {
         bytes32 metaCommitment,
         bytes32 resultsId
     ) external returns (bool);
-    function submitMetaChallenge(uint256 computeId, uint256 subJobId) external returns (bool);
+    function submitMetaChallenge(uint256 computeId, uint32 subJobId) external returns (bool);
     function isAllowlistedComputer(
         address computer
     ) external returns (bool);
